pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "sdxl-opt"
version = "0.1.0"
description = "Systematic SDXL compression and optimization study"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Rajat Gupta"}]

dependencies = [
    "torch>=2.1.0",
    "diffusers>=0.27.0",
    "transformers>=4.38.0",
    "accelerate>=0.27.0",
    "bitsandbytes>=0.43.0",
    "peft>=0.9.0",
    "safetensors",
    "open-clip-torch",
    "torchmetrics[image]",
    "litserve>=0.2.0",
    "pyyaml",
    "pandas",
    "matplotlib",
    "seaborn",
    "Pillow",
    "tqdm",
    "psutil",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "ipykernel", "jupyter"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
